+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.3] - 2023-06-29
+
+### Added
+
+- Add `gbasis` dependency for AIM moment calculation
+- Add `moments` module.
+
+## [Unreleased]
+
 ## [1.0.2] - 2023-06-19
 
 ### Fixed